# Dirent lookup speedup