--- a/meson_options.txt
+++ b/meson_options.txt
@@ -2,6 +2,8 @@ option('CLUSTER_CACHE_SIZE', type : 'string', value : '16',
   description : 'set cluster cache size to number (default:16)')
 option('DIRENT_CACHE_SIZE', type : 'string', value : '512',
   description : 'set dirent cache size to number (default:512)')
+option('DIRENT_LOOKUP_CACHE_SIZE', type : 'string', value : '1024',
+  description : 'set dirent lookup cache size to number (default:1024)')
 option('LZMA_MEMORY_SIZE', type : 'string', value : '128',
   description : 'set lzma uncompress memory in MB (default:128)')
 option('USE_MMAP', type: 'boolean', value: true,