--- a/src/config.h.in
+++ b/src/config.h.in
@@ -3,6 +3,8 @@
 
 #mesondefine DIRENT_CACHE_SIZE
 
+#mesondefine DIRENT_LOOKUP_CACHE_SIZE
+
 #mesondefine CLUSTER_CACHE_SIZE
 
 #mesondefine LZMA_MEMORY_SIZE