--- a/src/fileimpl.cpp
+++ b/src/fileimpl.cpp
@@ -62,8 +62,7 @@ offset_t readOffset(const Reader& reader, size_t idx)
       direntCache(envValue("ZIM_DIRENTCACHE", DIRENT_CACHE_SIZE)),
       direntCacheLock(PTHREAD_MUTEX_INITIALIZER),
       clusterCache(envValue("ZIM_CLUSTERCACHE", CLUSTER_CACHE_SIZE)),
-      cacheUncompressedCluster(envValue("ZIM_CACHEUNCOMPRESSEDCLUSTER", false)),
-      m_direntLookup(this)
+      cacheUncompressedCluster(envValue("ZIM_CACHEUNCOMPRESSEDCLUSTER", false))
   {
     log_trace("read file \"" << fname << '"');
 
@@ -165,6 +164,7 @@ offset_t readOffset(const Reader& reader, size_t idx)
 
       current += (len + 1);
     }
+    m_direntLookup.init(this, 1024);
   }