This PR addresses #410 & #424

The speed-up of dirent look-up is achieved via an auxiliary special purpose in-memory cache (new helper class `zim::NarrowDown`). It stores the full urls (ns+path) and indices of uniformly spaced dirents and is filled upon opening the ZIM file. Then the dirent look-up operation is performed in two stages:

1. Narrow down the range that might contain the dirent with the given full url via a binary search in that cache.
2. Perform binary search on dirents (with disc access as needed) in that narrower range.

Since the speed-up is due to avoiding expensive IO operations, the goal is to fit as much data as possible in the in-memory cache. This is achieved as follows:
1. A sorted vector container is used instead of a tree (`std::map`) data structure.
2. Actual URLs (ns+path) are approximated with shortened versions that allow to narrow-down the initial full range with almost the same efficiency except a few corner cases (see below for the explanation of this trick).
3. (The shortened) URLs are not stored as `std::string` objects (which have an overhead of a couple of pointers). Instead they are packed sequentially as C-style strings in a single array, and 4-byte offsets into that array are used.

The idea of shortening the URLs is illustrated on an example:

| Dirent #  |  Full URL (namespace concatenated with path)      |
|-------------|--------------------------------------------------------------------|
| ...           |         ...                                                                      |
| 1234      |  AList_of_Climate_Change_Initiatives                      |
| 1235      |  AList_of_Greta_Thunberg_speeches                      |
| ...           |         ...                                                                      |

The example above illustrates two consecutive entries from the wikipedia_en_climate_change_nopic_2020-01.zim ZIM file (though the index values are not real). If we were to use the item `#1234` in our cache the naive approach would be to store its full URL as is. However, let's imagine that the ZIM file also contained an item with full URL **AList_of_G**. Then it would have to reside between **AList_of_Climate_Change_Initiatives** and **AList_of_Greta_Thunberg_speeches**. So we can pretend that such an item exists and store in our cache the fictitious entry (**AList_of_G**, 1234.5). When we arrive at that entry during the range narrow-down step we must round the dirent index downward if it is going to be used as the lower bound of the range, and round it upward if it is going to be used as the upper bound of the range.

---

## Benchmark results

Since this optimization effectively adds 1024 additional (but static) dirent lookup entries, the benchmark includes two runs of the `master` version with different values for the `ZIM_DIRENTCACHE` setting - 512 (the default) and 1536 (default + 1024).

In summary, this optimization improves the performance of `zimcheck` more than simply increasing the dirent cache size does, and doesn't negatively impact the performance of `kiwix-serve`.

Data used for benchmarking:

| id |  ZIM file                                                                     |    size (MB) | article count | cluster count |
|----|---------------------------------------------------------------------|----------------|------------------|-------------------|
| 1 | wikipedia_en_climate_change_nopic_2020-01.zim  |  31    | 7646     | 51     |
| 2 | [wikipedia_hy_all_mini_2020-07.zim][2]                    |   560 | 509325 | 1518 |
| 3 | [wikipedia_hy_all_mini_2020-08.zim][1]                    |   563 | 509611 | 1526 |

---

### `zimcheck -A` benchmark

|   ZIM file id  | zimcheck -A runtime (master, DIRENTCACHE=512) | zimcheck -A runtime (master, DIRENTCACHE=1536) |  zimcheck -A runtime (this PR, DIRENTCACHE=512) | zimcheck -A runtime (this PR, DIRENTCACHE=1536) |
|-----|-----------------|-----------------|---------------|-----------------|
|  1  | 6.183s        | 5.881s        | 5.693s        |  5.531s       |
|  2  | 6m36.615s | 6m12.248s | 5m51.896s | 5m40.461s |
|  3  | 6m31.202s | 6m16.159s | 5m55.336s | 5m42.636s |

---
### `kiwix-serve` benchmark

The benchmark was performed using the [`benchmark_kiwix_serve`](https://github.com/veloman-yunkan/kiwix-dev-utils/blob/80562e5de15faba59dfec1c3a2da147a699825c9/benchmark_kiwix_serve) script as follows:

```
ZIM_DIRENTCACHE=512 ./benchmark_kiwix_serve master 1000 zimfiles/wikipedia_hy_all_mini_2020-08.zim
ZIM_DIRENTCACHE=1536 ./benchmark_kiwix_serve master 1000 zimfiles/wikipedia_hy_all_mini_2020-08.zim
ZIM_DIRENTCACHE=512 ./benchmark_kiwix_serve pr430 1000 zimfiles/wikipedia_hy_all_mini_2020-08.zim
```

That script starts a kiwix server and fetches from it about 1000 articles using `wget --recursive -l 1`.

|   ZIM file id     | wget runtime (master, DIRENTCACHE=512) | wget runtime (master, DIRENTCACHE=1536) |  wget runtime (this PR, DIRENTCACHE=512) |  wget runtime (this PR, DIRENTCACHE=1536) |
|--------------------|------------|-------------|------------|-------------|
|  3                    | 36.621s | 35.628s | 36.210s | 36.236s |

(It was verified that the output directories created by wget from both runs are identical, with the exception of `/random`)


[1]: http://download.kiwix.org/zim/wikipedia/wikipedia_hy_all_mini_2020-08.zim
[2]: http://download.kiwix.org/zim/wikipedia/wikipedia_hy_all_mini_2020-07.zim