@veloman-yunkan 2020-09-30
@kelson42 @mgautierfr 

Please see the benchmarking results of this draft PR before deciding how to proceed from here.