@mgautierfr 2020-09-30
> Please see the benchmarking results of this draft PR before deciding how to proceed from here.

The idea of splitting the PR is still valid. The first PR should not change the performance (but improve the code organization) and the second PR will work on performance (and let us discuss about different algorithm possibilities)

---

I'm a bit surprised by the benchmark results.
zimcheck loop over the dirent without using the `find` method (binary search) and this PR (the performance part) is reducing the range in which search. So the performance should be on the kiwix-serve side, not the zimcheck.