@veloman-yunkan 2020-09-30
> I'm a bit surprised by the benchmark results.
> zimcheck loop over the dirent without using the `find` method (binary search)

`zimcheck -A` includes URL validation which [checks that all internal links in the ZIM file are valid](https://github.com/openzim/zim-tools/blob/a46e5d935001e6c2ed90fd2094d6d1355621c912/src/zimcheck/checks.cpp#L182-L187).