@veloman-yunkan 2020-09-30
> To be complete, it would be really helpful to have your PR with 1536 Dirents in the cache.

Done