@kelson42 2020-09-30
@mgautierfr @veloman-yunkan Thank you. I'm still unsure about the memory impact of `DIRENTCACHE=1536`, but if this is not taking too much memory, we should better use it?