@veloman-yunkan 2020-09-30
> I'm still unsure about the memory impact of `DIRENTCACHE=1536`, but if this is not taking too much memory, we should better use it?

1536 is not a magic number and hardly it is the sweet spot for a typical runtime environment. Using a larger number will likely result in even more speed-up on average.