@mgautierfr 2020-09-30
In this PR you use a different approach of what I expose in #424.

Your implementation a 1024 items map and split the whole articles range into 1024 balanced subranges.
But the lookup for the subrange imply a `upper_bound` call (`O(log(1024))`, so `O(1)` as 1024 is const) which is a bit longer than a direct array indexing.

My proposition split the range in 256 sub-range (not balanced) (but allow subdividing sub-range into 256 sub-sub-range). It is a bit more complex to construct (as we build the subranges based on the content and so we have to parse the content). But the lookup is probably a bit faster as we do a simple array indexing. It also has the advantage to store less data (a 4 bytes offset per range) compared to yours (a 4 bytes offset and the dirent path).

It seems that your implementation is easier to implement (it is already implemented) and build the cache array faster. But it is potentially slower to lookup dirents once the cache is created. It seems my idea is better for long running process. (But yet again, I don't what is the real performance gain between the two ideas)

As you already said, your PR is draft and not optimized for memory usage. But I would like to know your point here. Does your algorithm is the "final" idea or it is just a first implementation of a array cache to have benchmark numbers and them we see what is best to implement ?