@mgautierfr 2020-09-30
>  How serious is that issue? 

I don't know :) 
Your current implementation is probably not a issue. The cache filling is just 1024 dirent reads.
While it takes time, it is not so much. (The actual implementation searching for start of namespaces with binary search may reads dozens of dirents, and my algo will probably need hundreds)

> Since kiwix-serve can be started with a library consisting of a lot of ZIM files, is it acceptable to initialize such a lookup data structure eagerly? Should we instead try to fill it on demand?

On kiwix-serve, I would say it is not a issue as we are in a long running process. The warmup time will not be so important, we will simply have to wait a bit longer before the server is up.
On a one shoot process, it may be more important, but it would be on only one zim.

On kiwix-serve, we open all zim files only to locate the xapian database to have multizim search. Else we would open zim file only on demand. I think we should have different strategy depending of the usage.
(Opening zim file without cache to locate xapian database or zimdump, Opening with cache for long process, ...). But this strategy selection will not be made here. For now, I think we can assume that the warmup time is not a real issue (if it is reasonable).