@kelson42 2020-10-09
@veloman-yunkan Hi, any news here? We wait to that PR to:
* complete speedup project on libzim
* make a new release of lizim and zimtools to benefit of the full speedup